+++ packages/atlas/src/views/viewer/CategoryView/CategoryView.tsx
@@ -46,20 +46,29 @@ export const CategoryView = () => {
 
   // const headTags = useHeadTags(currentCategory?.name)
 
-  // const { categoriesFeaturedVideos } = useCategoriesFeaturedVideos(currentCategory?.id || '')
+  // const { categoriesFeaturedVideos, loading: categoriesFeaturedVideosLoading } = useCategoriesFeaturedVideos(
+  //   currentCategory?.id || ''
+  // )
   // const videoHeroVideos = useVideoHeroVideos(categoriesFeaturedVideos)
 
   // return (
-  //   <VideoContentTemplate cta={['popular', 'new', 'home']}>
-  //     {headTags}
-  //     <VideoCategoryHero
-  //       header={{
-  //         title: currentCategory?.name ?? undefined,
-  //         icon: <CategoryIcon url={currentCategory?.iconUrl} color={cVar('colorTextStrong')} />,
-  //       }}
-  //       videos={videoHeroVideos}
-  //     />
-
+  // <VideoContentTemplate cta={['popular', 'new', 'home']}>
+  //   {headTags}
+  //   <VideoCategoryHero
+  //     loading={categoriesFeaturedVideosLoading}
+  //     category={{
+  //       title: currentCategory?.name ?? undefined,
+  //       icon: (
+  //         <CategoryIcon
+  //           url={currentCategory?.iconUrl}
+  //           color={videoHeroVideos.length ? cVar('colorTextStrong') : cVar('colorCoreBaseBlack')}
+  //         />
+  //       ),
+  //       color: currentCategory.color,
+  //       coverImgUrl: currentCategory.coverImgUrl,
+  //     }}
+  //     videos={videoHeroVideos}
+  //   />
   //     {!!categoriesFeaturedVideos?.length && (
   //       <>
   //         <TitleContainer>